% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_func.R
\name{dep_func}
\alias{dep_func}
\title{Template function to demonstrate lifecycle badges}
\usage{
dep_func(x, y, z, ...)
}
\arguments{
\item{x}{Placeholder argument}

\item{y}{Placeholder argument}

\item{z}{Placeholder argument}

\item{...}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Adding arguments through the
ellipsis is no longer supported.}
}
\value{
Nothing, called for printing side-effect.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
