% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expe_func.R
\name{expe_func}
\alias{expe_func}
\title{Template function to demonstrate lifecycle badges}
\usage{
expe_func(x, y, z, ...)
}
\arguments{
\item{x}{Placeholder argument}

\item{y}{Placeholder argument}

\item{z}{Placeholder argument}

\item{...}{Extra arguments to be passed to function}
}
\value{
Nothing, called for printing side-effect.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
foo()
}
