% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycling-package.R
\docType{package}
\name{lifecycling-package}
\alias{lifecycling}
\alias{lifecycling-package}
\title{lifecycling: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
