% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_func.R
\name{sup_func}
\alias{sup_func}
\title{Template function to demonstrate lifecycle badges}
\usage{
sup_func(x, y, z, ...)
}
\arguments{
\item{x}{Placeholder argument}

\item{y}{Placeholder argument}

\item{z}{Placeholder argument}

\item{...}{Extra arguments to be passed to function}
}
\value{
Nothing, called for printing side-effect.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
\examples{
sup_func()
}
