% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycling-package.R
\docType{package}
\name{lifecycling-package}
\alias{lifecycling}
\alias{lifecycling-package}
\title{lifecycling: Explains how semantic versioning and lifecycles in R}
\description{
This package acts as a demonstration on how to use semantic versioning and lifecycle badges from different badge frameworks to communicate open-source software development with the community.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joshwlambert/lifecycling}
  \item \url{https://joshwlambert.github.io/lifecycling/}
  \item Report bugs at \url{https://github.com/joshwlambert/lifecycling/issues}
}

}
\author{
\strong{Maintainer}: Joshua W. Lambert \email{joshua.lambert@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID}) [copyright holder]

}
\keyword{internal}
